refactor: rename `manager.rs` -> `provider_manager.rs`
