+++ packages/desktop/src/providers/mod.rs
@@ -5,10 +5,10 @@ pub mod host;
 pub mod ip;
 #[cfg(windows)]
 pub mod komorebi;
-pub mod manager;
 pub mod memory;
 pub mod network;
 pub mod provider;
+pub mod provider_manager;
 pub mod provider_ref;
 pub mod variables;
 pub mod weather;
