+++ packages/desktop/src/providers/provider_manager.rs
